fastapi_x_react/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "authlib>=1.6.8",
    "bcrypt==4.3.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
